Better makefile, long long int support for ft_atoi

+++ 0-libft/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: pdong <pdong@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/05 14:19:53 by pdong         #+#    #+#                 */
-/*   Updated: 2022/10/05 14:50:40 by pdong         ########   odam.nl         */
+/*   Updated: 2022/11/15 15:46:31 by pdong         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 // #include <stdio.h>
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
-	int	i;
-	int	neg;
-	int	ans;
+	int			i;
+	int			neg;
+	long long	ans;
 
 	i = 0;
 	neg = 1;
@@ -40,8 +40,8 @@ int	ft_atoi(const char *str)
 
 // int main () 
 // {
-//    const char str[50] = "  \n\f -1234a567  ";
-//    int a = ft_atoi(str);
-//    printf("%d\n", a);
+//    const char str[50] = "  \n\f -1234123123123a567  ";
+//    long long a = ft_atoi(str);
+//    printf("%lld\n", a);
 //    return(0);
 // }
